requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
dj-database-url==2.1.0
Django==3.2.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.12.2
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
whitenoise==6.6.0
